Refactored Product model, minor refactoring to User model, add Category and ProductOrder models

+++ src/main/java/com/example/springblog/repo/ProductRepository.java
@@ -4,10 +4,10 @@ import com.example.springblog.models.Product;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Product, Integer> {
+public interface ProductRepository extends CrudRepository<Product, Long> {
     @Query("from Product a where a.user=?1")
-    Product findUserId(int id);
+    Product findUserId(long id);
     @Query("delete from Product a where a.id=?1")
-    Product deleteById(int id);
+    Product deleteById(long id);
     Product findByTitle(String title);
 }
